refactor(sandbox): navigation-patterns 専用コンポーネントを features/ へ移動
presentation-launcher と presentation-sample は navigation-patterns 機能
（on-root / in-tab の2ルートツリー）からのみ利用されており、共有 UI では
ないため src/features/navigation-patterns/ へ co-location する。

- src/components/ は「複数機能から使われる共有 UI」の置き場として再定義
  （bulletproof-react や Expo 公式テンプレートと同じ整理）
- PressableCard は Card/CardShape と一体の variant ファミリーであり
  navigation-patterns のドメイン知識を持たないため components/card/ に残す
- .po の差分は移動に伴う origins（参照パス）の更新のみ
