+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/form-sheet/in-tab.tsx
@@ -2,7 +2,7 @@ import type { ReactElement } from "react";
 import { View } from "react-native";
 import { Stack } from "expo-router";
 import { Trans, useLingui } from "@lingui/react/macro";
-import { PresentationSampleScreen } from "../../../../../components/presentation-sample/PresentationSampleScreen";
+import { PresentationSampleScreen } from "../../../../../features/navigation-patterns/presentation-sample/PresentationSampleScreen";
 
 export default function FormSheetSample(): ReactElement {
   const { t } = useLingui();
