+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/full-screen-modal/in-tab-slide-from-bottom.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from "react";
 import { Stack } from "expo-router";
 import { Trans, useLingui } from "@lingui/react/macro";
-import { PresentationSampleScreen } from "../../../../../components/presentation-sample/PresentationSampleScreen";
+import { PresentationSampleScreen } from "../../../../../features/navigation-patterns/presentation-sample/PresentationSampleScreen";
 
 export default function FullScreenModalSlideFromBottomSample(): ReactElement {
   const { t } = useLingui();
