+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/modal/index.tsx
@@ -4,7 +4,7 @@ import { Trans, useLingui } from "@lingui/react/macro";
 import {
   type LauncherSection,
   PresentationLauncherList,
-} from "../../../../../components/presentation-launcher/PresentationLauncherList";
+} from "../../../../../features/navigation-patterns/presentation-launcher/PresentationLauncherList";
 
 export default function ModalIntermediate(): ReactElement {
   const { t } = useLingui();
